Rollback changes of namespaces
