+++ Source/EasyNetQ.Tests.Tasks/Tasks/TestScheduledMessages.cs
@@ -2,7 +2,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using EasyNetQ.Producer;
-using EasyNetQ.Scheduler;
+using EasyNetQ.Scheduling;
 using Net.CommandLine;
 using Serilog;
