+++ Source/EasyNetQ/IBus.cs
@@ -1,6 +1,6 @@
 using System;
 using EasyNetQ.Producer;
-using EasyNetQ.Scheduler;
+using EasyNetQ.Scheduling;
 
 namespace EasyNetQ
 {
