+++ Source/EasyNetQ/Scheduling/DelayedExchangeScheduler.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using EasyNetQ.Topology;
 
-namespace EasyNetQ.Scheduler
+namespace EasyNetQ.Scheduling
 {
     public class DelayedExchangeScheduler : IScheduler
     {
