+++ Source/EasyNetQ/Scheduling/SchedulerExtensions.cs
@@ -1,29 +1,33 @@
 ﻿using System;
 using System.Threading;
-using System.Threading.Tasks;
 
-namespace EasyNetQ.Scheduler
+namespace EasyNetQ.Scheduling
 {
-    /// <summary>
-    /// Provides a simple Publish API to schedule a message to be published at some time in the future.
-    /// </summary>
-    //TODO Need support of non-generic overloads
-    public interface IScheduler
+    public static class SchedulerExtensions
     {
         /// <summary>
         /// Schedule a message to be published at some time in the future.
         /// This required the EasyNetQ.Scheduler service to be running.
         /// </summary>
         /// <typeparam name="T">The message type</typeparam>
+        /// <param name="scheduler">The scheduler instance</param>
         /// <param name="delay">The delay for message to publish in future</param>
         /// <param name="message">The message to response with</param>
         /// <param name="topic">The topic string</param>
         /// <param name="cancellationToken">The cancellation token</param>
-        Task FuturePublishAsync<T>(
-            T message,
+        public static void FuturePublish<T>(
+            this IScheduler scheduler,
+            T message, 
             TimeSpan delay,
-            string topic = null,
+            string topic = null, 
             CancellationToken cancellationToken = default
-        );
+        )
+        {
+            Preconditions.CheckNotNull(scheduler, "scheduler");
+            
+            scheduler.FuturePublishAsync(message, delay, topic, cancellationToken)
+                .GetAwaiter()
+                .GetResult();
+        }
     }
 }
